Initial code.
Been working on this all day. I've changed it a bit from the default arduino project format by splitting it into different files. I've also opted for using my own main instead of the auto-generated one. 

Right now it does not take configuration commands, but it does have a configuration class that can access where the configuration is stored. Packets are also in a struct to make it easier to send/recv packets. Because of this, the main code isn't super difficult since it's all hidden behind class implementations.